requirements.txt
nose==1.3.7
pynsq==0.6.8
PyYAML==3.11
six==1.10.0
statsd==3.2.1
tornado==4.3
uritools==0.7.0
wsgiref==0.1.2
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
singledispatch==3.4.0.3
certifi==2015.11.20.1
MySQL-python==1.2.5
